requirements.txt
# Pinned for Python 3.10
fastapi==0.115.12
uvicorn[standard]==0.34.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
tensorflow==2.15.1
keras==2.15.0
mysql-connector-python==8.4.0
python-dotenv==1.0.1
matplotlib==3.9.4
seaborn==0.13.2
joblib==1.4.2
